+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.1.0
+* TCP sockets support implemented (breaking changes).
+* ServeMeClient constructor now takes single ServeMeSocket argument.
+* Future<ServeMeClient> ServeMe.connect() method added allowing to create WebSocket or TCP client connections (with the same functionality as server client connections).
+
 ## v1.0.2
 * Bugfix: ESC key press caused ServeMe to crash on Linux